+++ option-t.d.ts
@@ -40,6 +40,9 @@ declare module 'option-t' {
         unwrap(): T;
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
+        and<U>(optb: Option<U>): Option<U>;
+        or(optb: Option<T>): Option<T>;
+        orElse(fn: () => Option<T>): Option<T>;
         drop(): void;
     }
 
@@ -49,6 +52,9 @@ declare module 'option-t' {
         unwrap(): T;
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
+        and<U>(optb: Option<U>): Option<U>;
+        or(optb: Option<T>): Option<T>;
+        orElse(fn: () => Option<T>): Option<T>;
         drop(): void;
     }
 
@@ -58,6 +64,9 @@ declare module 'option-t' {
         unwrap(): T;
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
+        and<U>(optb: Option<U>): Option<U>;
+        or(optb: Option<T>): Option<T>;
+        orElse(fn: () => Option<T>): Option<T>;
         drop(): void;
     }
 }